Refactoring/clean up of import data classes
