+++ modules/admin/app/services/ingest/WSIngestService.scala
@@ -17,77 +17,15 @@ import defines.{ContentTypes, EntityType}
 import javax.inject.Inject
 import play.api.http.HeaderNames
 import play.api.libs.Files.SingletonTemporaryFileCreator
-import play.api.libs.json.JsonConfiguration.Aux
-import play.api.libs.json.JsonNaming.SnakeCase
 import play.api.libs.json._
 import play.api.libs.ws.{BodyWritable, SourceBody, WSClient}
 import play.api.{Configuration, Logger}
-import services.data.{Constants, ValidationError}
+import services.data.Constants
 import services.redirects.MovedPageLookup
 import services.search.{SearchConstants, SearchIndexMediator}
 import services.storage.FileStorage
 
 import scala.concurrent.{ExecutionContext, Future}
-
-
-// A result from the import endpoint
-sealed trait IngestResult
-
-object IngestResult {
-  implicit val reads: Reads[IngestResult] = Reads { json =>
-    json.validate[ValidationError]
-      .map(e => ErrorLog("Validation error", e.toString))
-      .orElse(json.validate[ErrorLog])
-      .orElse(json.validate[ImportLog])
-      .orElse(json.validate[SyncLog])
-  }
-  implicit val writes: Writes[IngestResult] = Writes {
-    case i: ImportLog => Json.toJson(i)(ImportLog.format)
-    case i: SyncLog => Json.toJson(i)(SyncLog.format)
-    case i: ErrorLog => Json.toJson(i)(ErrorLog.format)
-  }
-  implicit val format: Format[IngestResult] = Format(reads, writes)
-}
-
-// The result of a regular import
-case class ImportLog(
-  createdKeys: Map[String, Seq[String]] = Map.empty,
-  created: Int = 0,
-  updatedKeys: Map[String, Seq[String]] = Map.empty,
-  updated: Int = 0,
-  unchangedKeys: Map[String, Seq[String]] = Map.empty,
-  unchanged: Int = 0,
-  message: Option[String] = None,
-  event: Option[String] = None,
-  errors: Map[String, String] = Map.empty,
-) extends IngestResult {
-  def hasDoneWork: Boolean = createdKeys.nonEmpty || updatedKeys.nonEmpty
-}
-
-object ImportLog {
-  implicit val config: Aux[Json.MacroOptions] = JsonConfiguration(SnakeCase)
-  implicit val format: Format[ImportLog] = Json.format[ImportLog]
-}
-
-// The result of an EAD synchronisation, which incorporates an import
-case class SyncLog(
-  deleted: Seq[String],
-  created: Seq[String],
-  moved: Map[String, String],
-  log: ImportLog
-) extends IngestResult
-
-object SyncLog {
-  implicit val format: Format[SyncLog] = Json.format[SyncLog]
-}
-
-// An import error we can understand, e.g. not a crash!
-case class ErrorLog(error: String, details: String) extends IngestResult
-
-object ErrorLog {
-  implicit val format: Format[ErrorLog] = Json.format[ErrorLog]
-}
-
 
 
 /**
@@ -100,7 +38,6 @@ case class WSIngestService @Inject()(
   searchIndexer: SearchIndexMediator,
   pageRelocator: MovedPageLookup,
   fileStorage: FileStorage,
-  importLogger: ImportLogService
 )(implicit actorSystem: ActorSystem, mat: Materializer) extends IngestService {
 
   import services.ingest.IngestService._
@@ -144,28 +81,12 @@ case class WSIngestService @Inject()(
   override def storeManifestAndLog(job: IngestJob, res: IngestResult): Future[URI] = {
     import play.api.libs.json._
 
-    implicit val payloadHandleWrites: Writes[PayloadHandle] = Writes {
-      case FilePayload(f) => Json.toJson(f.map(_.toAbsolutePath.toString))
-      case UrlMapPayload(urls) => Json.toJson(urls.mapValues(_.toString))
-    }
-
     val data = Json.obj(
       "id" -> job.id,
       "type" -> job.data.dataType,
-      "params" -> Json.obj(
-        "scope" -> job.data.params.scope,
-        "lang" -> job.data.params.lang,
-        "fonds" -> job.data.params.fonds,
-        "allow-update" -> job.data.params.allowUpdate,
-        "handler" -> job.data.params.handler,
-        "importer" -> job.data.params.importer,
-        "excludes" -> job.data.params.excludes,
-        "properties" -> job.data.params.properties.toString,
-        "commit" -> job.data.params.commit
-      ),
+      "params" -> job.data.params,
       "user" -> job.data.user.toOption,
       "stats" -> res,
-      "data" -> job.data.params.data
     )
 
     val bytes = Source.single(ByteString.fromArray(
